seccompiler: support riscv64 target
Firecracker creates seccomp filters at build time, and fails if there is
no support for the targeted architecture. This commit adds support for
the RISC-V architecture in the seccompiler crate.

For now, we do not supply any riscv64-specific JSON file in
`resources/seccomp`, adhering to the default one, since RISC-V support
is experimental.

Signed-off-by: Dimitris Charisis <dchar@cslab.ece.ntua.gr>